// src/modules/order/order.enum.ts
export enum ORDER_TYPE {
  STANDARD = 'STANDARD',
  QUICK = 'QUICK',
}

export enum SERVICE_TYPE {
  TRAILOR_TRACTOR_TRUCK = 'TRAILOR_TRACTOR_TRUCK',
  NORMAL_TRUCK_VAN = 'NORMAL_TRUCK_VAN',
  NON_MOTORIZED = 'NON_MOTORIZED_VEHICLE',
  CONCATENATED_GOODS = 'CONCATENATED_GOODS',
  CONTRACT_CAR = 'CONTRACT_CAR',
}

export enum CONTAINER_TYPE {
  NO = '',
  DC = 'DC',
  RF = 'RF',
  HC = 'HC',
  HR = 'HR',
  OT = 'OT',
  FR = 'FR',
  TANK = 'TANK',
}

export enum CONTAINER_SIZE {
  NO = '',
  SIZE_2x20 = '2x20',
  SIZE_20 = '20',
  SIZE_40 = '40',
  SIZE_45 = '45',
}

export enum CANCELED_BY {
  CUSTOM_CANCEL,
  DRIVER_CANCEL,
}
export enum CARGO_TYPE {
  NO = '',
  STANDARD = 'STANDARD',
  DANGEROUS = 'DANGEROUS',
  SPECIAL = 'SPECIAL',
}
export enum PACKAGE_SIZE {
  PACKAGE_SIZE_1000_1200 = '1000x1200',
  PACKAGE_SIZE_1100_1100 = '1100x1100',
  PACKAGE_SIZE_1016_1219 = '1016x1219',
  PACKAGE_SIZE_1165_1165 = '1165x1165',
  PACKAGE_SIZE_1067_1067 = '1067x1067',
  PACKAGE_SIZE_800_1200 = '800x1200',
  PACKAGE_SIZE_OTHERS = 'Others',
}
export enum ORDER_STATUS {
  CREATED = 'CREATED',
  VERIFIED = 'VERIFIED',
  ASSIGNING = 'ASSIGNING',
  ASSIGNED = 'ASSIGNED',
  DISPATCHED = 'DISPATCHED',
  PICKING = 'PICKING',
  PICK_ARRIVED = 'PICK_ARRIVED',
  PICKUPCODE_INPUTED = 'PICKUPCODE_INPUTED',
  DELIVERING = 'DELIVERING',
  DELIVERYCODE_INPUTED = 'DELIVERYCODE_INPUTED',
  DELIVERED = 'DELIVERED',
  CANCELING = 'CANCELING',
  CANCELED = 'CANCELED',
  CUSTCANCEL = 'CUSTCANCEL',
  DRIVERCANCEL = 'DRIVERCANCEL',
}

export enum TRUCK_SPECIAL_TYPE {
  NO = '',
  NORMAL = 'NORMAL',
  FREEZING = 'FREEZING',
}

export enum ACTIONS_MODE {
  CREATE = 'create',
  EDIT = 'edit',
}

export enum ADDITIONAL_PRICE {
  ADJUSTMENT_AMOUNT = 'ADJUSTMENT_AMOUNT',
  FORKLIFT_COST = 'FORKLIFT_COST',
  LABOUR_COST = 'LABOUR_COST',
  ROAD_FEE = 'ROAD_FEE',
  WAITING_FEE = 'WAITING_FEE',
  INCENTIVES = 'INCENTIVES',
  OTHERS = 'OTHERS',
}